# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_emzola_issuetracker_internal_model.IssuesAssignee:
    properties:
      assignee_id:
        type: integer
      assignee_name:
        type: string
      issues_assigned:
        type: integer
    type: object
  github_com_emzola_issuetracker_internal_model.IssuesPriority:
    properties:
      issue_priority:
        type: string
      issues_count:
        type: integer
    type: object
  github_com_emzola_issuetracker_internal_model.IssuesReporter:
    properties:
      issues_reported:
        type: integer
      reporter_id:
        type: integer
      reporter_name:
        type: string
    type: object
  github_com_emzola_issuetracker_internal_model.IssuesStatus:
    properties:
      issue_status:
        type: string
      issues_count:
        type: integer
    type: object
  github_com_emzola_issuetracker_internal_model.IssuesTargetDate:
    properties:
      issue_title:
        type: string
      target_resolution_date:
        type: string
    type: object
...
host: https://issuetracker-api-dev.fl0.io
info:
  contact:
    email: emma.idika@yahoo.com
    name: API Support
  description: This is an API service for an issue tracker.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Issue Tracker API
  version: 1.0.0
paths:
...
  /v1/issuesreport/assignee:
    get:
      description: This endpoint gets report of issue assignees for a project
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for project_id
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.IssuesAssignee'
            type: array
        "500":
          description: Internal Server Error
      summary: Get report of issue assignees for a project
      tags:
      - issuesreport
  /v1/issuesreport/date:
    get:
      description: This endpoint gets report of issue target date for a project
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for project_id
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.IssuesTargetDate'
            type: array
        "500":
          description: Internal Server Error
      summary: Get report of issues target date for a project
      tags:
      - issuesreport
  /v1/issuesreport/priority:
    get:
      description: This endpoint gets report of issues priority level for a project
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for project_id
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.IssuesPriority'
            type: array
        "500":
          description: Internal Server Error
      summary: Get report of issues priority level for a project
      tags:
      - issuesreport
  /v1/issuesreport/reporter:
    get:
      description: This endpoint gets report of issues reporter for a project
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for project_id
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.IssuesReporter'
            type: array
        "500":
          description: Internal Server Error
      summary: Get report of issues reporter for a project
      tags:
      - issuesreport
  /v1/issuesreport/status:
    get:
      description: This endpoint gets report of issue status for a project
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for project_id
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.IssuesStatus'
            type: array
        "500":
          description: Internal Server Error
      summary: Get report of issue status for a project
      tags:
      - issuesreport
...
swagger: "2.0"